--- students/gotem888/hw-12_searchEngine/css/style.css
@@ -1,299 +1,0 @@
-@property --rotate {
-  syntax: "<angle>";
-  initial-value: 132deg;
-  inherits: false;
-}
-:root {
-  --form-height: 40vh;
-  --form-width: calc(var(--form-height) / 1.5);
-  --blue: #3498db;
-}
-@keyframes --move {
-  from {
-    z-index: 1;
-    transform: scale(1);
-  }
-  to {
-    z-index: 2;
-    transform: scale(3);
-  }
-}
-@keyframes --spin {
-  0% {
-    --rotate: 0deg;
-  }
-  100% {
-    --rotate: 360deg;
-  }
-}
-
-body {
-  font-family: "Roboto", sans-serif;
-  margin: 0;
-  min-height: 100vh;
-  background: #191c29;
-}
-
-.wrapper {
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-between;
-  align-items: flex-start;
-  gap: 30px;
-  padding: 40px 20px;
-}
-.form {
-  background: rgba(25, 28, 41, 1);
-  display: flex;
-  flex-flow: row wrap;
-  position: relative;
-  width: 300px;
-  height: var(--card-height);
-  padding: 40px 15px;
-  border-radius: 6px;
-  font-size: 18px;
-  color: rgb(88 199 250 / 100%);
-  cursor: pointer;
-  font-family: Poppins;
-  transition: 2s;
-}
-
-.form:hover {
-  /* transition: 2s; */
-  background-color: #202436;
-  color: rgb(88 199 250 / 100%);
-  box-shadow: 0 0 3px 1px #2e6e80, 0 0 5px 2px #5ddcff;
-}
-.form:hover:before,
-.form:hover:after {
-  animation: none;
-  opacity: 0;
-  transition: 2s;
-}
-
-.form::before {
-  content: "";
-  width: 104%;
-  height: 102%;
-  border-radius: 8px;
-  background-image: linear-gradient(
-    var(--rotate),
-    #5ddcff,
-    #3c67e3 70%,
-    #4e00c2
-  );
-  position: absolute;
-  z-index: -1;
-  top: -1%;
-  left: -2%;
-  animation: --spin 2.5s linear infinite;
-}
-
-.form::after {
-  position: absolute;
-  content: "";
-  top: calc(var(--form-height) / 6);
-  left: 0;
-  right: 0;
-  z-index: -1;
-  height: 100%;
-  width: 100%;
-  margin: 0 auto;
-  transform: scale(0.8);
-  filter: blur(calc(var(--form-height) / 5));
-  background-image: linear-gradient(
-    var(--rotate),
-    #5ddcff,
-    #3c67e3 43%,
-    #4e00c2,
-    rgb(233, 47, 47)
-  );
-  opacity: 1;
-  transition: opacity 2s;
-  animation: --spin 2.5s linear infinite;
-}
-
-.form .form__item {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin-bottom: 20px;
-}
-
-.form__item--text input {
-  width: 100%;
-  padding: 10px 0;
-  font-size: 20px;
-  color: #fff;
-  margin-bottom: 30px;
-  border: none;
-  border-bottom: 1px solid #fff;
-  outline: none;
-  background: transparent;
-}
-.form__item--text p {
-  display: inline-block;
-  color: red;
-}
-
-.form__item--text label {
-  position: absolute;
-  top: 50px;
-  left: 20px;
-  padding: 10px 0;
-  font-size: 20px;
-  color: #fff;
-  pointer-events: none;
-  transition: 0.5s;
-}
-.form__item--text input:focus ~ label p#here,
-.form__item--text input:valid ~ label p#here {
-  visibility: hidden;
-  transition: 0.5s;
-}
-
-.form__item--text input:focus ~ label,
-.form__item--text input:valid ~ label {
-  top: 20px;
-  color: #03e9f4;
-  font-size: 12px;
-}
-
-.form label {
-  cursor: pointer;
-}
-
-.form .form__item.form__item--radio {
-  flex-direction: row;
-  width: 70%;
-}
-div.radio {
-  margin-bottom: 30px;
-  width: 100%;
-}
-.form button.form__item {
-  font-family: Poppins;
-  font-size: 18px;
-  cursor: pointer;
-  align-items: center;
-  justify-content: center;
-  background-color: transparent;
-  border: 0px solid rgb(183, 187, 183);
-  border-radius: 3px;
-  margin: 0 30px 20px;
-  height: 40px;
-  transition: all 0.3s ease-in-out;
-  color: rgb(247, 12, 12);
-}
-
-.form button.form__item[type="submit"] {
-  border-color: #3496d6;
-  color: rgb(223, 223, 223);
-  transition: all 0.3s ease-in-out;
-}
-
-.form:hover button.form__item[type="submit"] {
-  box-shadow: 0 0 10px 0 #3498db inset, 0 0 10px 4px #3498db;
-}
-.form button.form__item[type="submit"]:active {
-  box-shadow: 0 0 30px 30px #2d89c7 inset;
-  color: #202436;
-  transition: 0.3s;
-}
-
-.form:hover button.form__item[type="reset"] {
-  box-shadow: 0 0 10px 0 #cc2525 inset, 0 0 10px 4px #e71313;
-  color: rgb(255, 6, 6);
-}
-.form button.form__item[type="reset"]:active {
-  box-shadow: 0px 0px 30px 30px rgb(184, 7, 7) inset;
-  color: #fff;
-  transition: 0.3s;
-}
-
-.wrapper__content {
-  width: calc(100% - 300px);
-  min-height: 500px;
-  padding: 20px 15px;
-  background: #191c29;
-  border: 0 #eee solid;
-  /* box-shadow: 0px 0px 8px 0px #fff; */
-}
-
-.loading {
-  text-align: center;
-}
-
-.content {
-  display: flex;
-  flex-flow: row wrap;
-  align-items: center;
-  justify-content: space-around;
-  width: 100%;
-  height: 100%;
-  gap: 10px;
-}
-
-.card {
-  display: flex;
-  position: relative;
-  height: 100%;
-  width: 160px;
-  border: 1px transparent solid;
-  border-radius: 3px;
-  box-shadow: 0px 0px 15px 0px rgba(148, 143, 143, 1);
-  cursor: pointer;
-  overflow: hidden;
-  transition: all 2s ease-out;
-}
-
-.card__img {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-  border-radius: 3px;
-}
-.card:hover {
-  animation: --move 2s linear alternate forwards;
-}
-
-.card__title {
-  font-family: Poppins;
-  font-size: 14px;
-  position: absolute;
-  height: auto;
-  bottom: 0;
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-  width: 100%;
-  text-align: center;
-  background-color: rgb(180, 12, 12);
-  color: white;
-  padding: 5px 0px;
-  margin: 0;
-  transition: height 0.2s ease-in-out;
-}
-.card:hover .card__title {
-  background: linear-gradient(
-    to top,
-    #ffffff 45.4%,
-    rgba(255, 255, 255, 0) 90%
-  );
-  color: rgb(180, 12, 12);
-  font-weight: bold;
-  transition: 1s;
-}
-.card:hover .content {
-  top: 50px;
-}
-.posCalibrate {
-  position: fixed;
-  top: 20%;
-}
-.wrapper__filters {
-  position: relative;
-  width: 300px;
-  height: auto;
-  margin: 0 50px 0 0;
-}
