--- students/gotem888/hw-12_searchEngine/scripts/scripts.js
@@ -1,375 +1,0 @@
-const PLUGINS = [
-  {
-    group_type: "instruments",
-    headline: "Arturia ARP2600 V2",
-    image: {
-      filename: "https://i.ibb.co/qW0yMXK/Arturia-ARP2600-V2.png",
-      alt: "ARP2600 V2",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Arturia CS-80",
-    image: {
-      filename: "https://i.ibb.co/k9QrQB1/Arturia-CS-80.png",
-      alt: "CS-80",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Waves Abbey Road Plates",
-    image: {
-      filename: "https://i.ibb.co/6BjK1hp/Waves-Abbey-Road-Plates-Stereo.png",
-      alt: "Abbey Road Plates",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Arturia Matrix V12",
-    image: {
-      filename: "https://i.ibb.co/Rv1dhmM/Arturia-Matrix-V-12.png",
-      alt: "Matrix V12",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Arturia Mini V",
-    image: {
-      filename: "https://i.ibb.co/Vgc40C7/Arturia-mini-V.png",
-      alt: "Mini V",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Arturia Modular",
-    image: {
-      filename: "https://i.ibb.co/6FYyRgW/Arturia-Modular.png",
-      alt: "Modular",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Arturia Oberheim SEM",
-    image: {
-      filename: "https://i.ibb.co/nwrwCff/Arturia-Oberheim-SEM.png",
-      alt: "Oberheim SEM",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Pure Mix SUGAR",
-    image: {
-      filename: "https://i.ibb.co/yQJsY6c/Pure-Mix-SUGAR.png",
-      alt: "SUGAR",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Arturia Prophet-V",
-    image: {
-      filename: "https://i.ibb.co/ygyNjbb/Arturia-Prophet-V.png",
-      alt: "Prophet-V",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Audiorealism ABL Pro",
-    image: {
-      filename: "https://i.ibb.co/72Skjjt/Audiorealism-ABL-Pro.png",
-      alt: "ABL Pro",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Image Line Hardcore",
-    image: {
-      filename: "https://i.ibb.co/qkxCKwy/Image-Line-Hardcore.png",
-      alt: "Hardcore",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "D16 Group LuSH-101",
-    image: {
-      filename: "https://i.ibb.co/PGk29M0/D-16-Group-Lu-SH-101.png",
-      alt: "LuSH-101",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Dmitry Sches Diversion",
-    image: {
-      filename: "https://i.ibb.co/Jxp583V/Dmitry-Sches-Diversion.png",
-      alt: "Diversion",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "FAW Circle2",
-    image: {
-      filename: "https://i.ibb.co/ZWpDZ22/FAW-Circle2.png",
-      alt: "Circle2",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Image Line GMS",
-    image: {
-      filename: "https://i.ibb.co/sKPDcF4/image-line-GMS.png",
-      alt: "GMS",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "HOFA IQ Reverb",
-    image: {
-      filename: "https://i.ibb.co/9bwFWCZ/HOFA-IQ-Reverb.png",
-      alt: "IQ Reverb",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Image Line Harmor",
-    image: {
-      filename: "https://i.ibb.co/1ry5NWh/image-line-Harmor.png",
-      alt: "Harmor",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "OUTPUT Movement",
-    image: {
-      filename: "https://i.ibb.co/Wz6kFZX/OUTPUT-Muvment.png",
-      alt: "Movement",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Image Line Sawer",
-    image: {
-      filename: "https://i.ibb.co/Y0RgVpB/image-line-Sawer.png",
-      alt: "Sawer",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Native Instruments Massive",
-    image: {
-      filename: "https://i.ibb.co/g791Zvy/Native-Instruments-Massive.png",
-      alt: "Massive",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Reveal Sound Spire",
-    image: {
-      filename: "https://i.ibb.co/sK01gz9/Reveal-Sound-Spire.png",
-      alt: "Spire",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Rob Papen Predator 3",
-    image: {
-      filename: "https://i.ibb.co/cCZ8dpJ/Rob-Papen-Predator3.png",
-      alt: "Predator 3",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Roland TB-303",
-    image: {
-      filename: "https://i.ibb.co/BVVPhbK/Roland-TB-303.png",
-      alt: "TB-303",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Roland TR-808",
-    image: {
-      filename: "https://i.ibb.co/qdYvkvH/Roland-TR-808.png",
-      alt: "TR-808",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Softube Parallels",
-    image: {
-      filename: "https://i.ibb.co/QDwfjBt/Softube-Parallels.png",
-      alt: "Parallels",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Waves Codex",
-    image: {
-      filename: "https://i.ibb.co/qBVy2sD/Waves-Codex.png",
-      alt: "Codex",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Waves Element",
-    image: {
-      filename: "https://i.ibb.co/RBJwvCM/Waves-Element.png",
-      alt: "Element",
-    },
-  },
-  {
-    group_type: "instruments",
-    headline: "Xfer Serum",
-    image: {
-      filename: "https://i.ibb.co/CWwcqyJ/Xfer-Serum.png",
-      alt: "Serum",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Baby Audio Spaced Out",
-    image: {
-      filename: "https://i.ibb.co/yfMVfKq/BABY-AUDIO-Spaced-Out.png",
-      alt: "Spaced Out",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "D16 Group Antresol",
-    image: {
-      filename: "https://i.ibb.co/K0vtzLR/D16-Group-Antresol.png",
-      alt: "Antresol",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "D16 Group Sigmund",
-    image: {
-      filename: "https://i.ibb.co/fxSxpvn/D16-Group-Sigmund.png",
-      alt: "Sigmund",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "D16 Group Spacerek",
-    image: {
-      filename: "https://i.ibb.co/SrpycS3/D16-Group-Spacerek.png",
-      alt: "Spacerek",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "D16 Group Toraverb 2",
-    image: {
-      filename: "https://i.ibb.co/HrPY2n2/D16-Group-Toraverb2.png",
-      alt: "Toraverb 2",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Fav Filter Pro-Q3",
-    image: {
-      filename: "https://i.ibb.co/r50qW4X/Fab-Filter-Pro-Q3.png",
-      alt: "Pro-Q3",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "HOFA IQ Analyser",
-    image: {
-      filename: "https://i.ibb.co/9hNXZC6/HOFA-IQ-Analyser.png",
-      alt: "IQ Analyser",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Image Line Convolver",
-    image: {
-      filename: "https://i.ibb.co/yqbybrs/Image-Line-Convolver.png",
-      alt: "Convolver",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Image Line Effector",
-    image: {
-      filename: "https://i.ibb.co/X35pTpd/Image-Line-Effector.png",
-      alt: "Effector",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "Lifeline Expanse",
-    image: {
-      filename: "https://i.ibb.co/DkRRQS5/Lifeline-Expanse.png",
-      alt: "Expanse",
-    },
-  },
-  {
-    group_type: "effects",
-    headline: "MURDA Melodies",
-    image: {
-      filename: "https://i.ibb.co/rQzn3Y4/MURDA-Melodies.png",
-      alt: "MURDA Melodies",
-    },
-  },
-];
-const createCardContent = (object) => {
-  const { headline, image } = object;
-
-  const card = `<div class="card">
-      <img class="card__img" src="${image.filename}" alt="${image.alt}">
-      <p class="card__title">${headline}</p>
-      </div>`;
-
-  return card;
-};
-
-const generateContent = (array) => {
-  const content = document.querySelector("#content");
-
-  content.innerHTML = "";
-
-  let data = "";
-
-  array.forEach((element) => {
-    data += createCardContent(element);
-  });
-
-  if (!data) data = "No data";
-
-  content.innerHTML = data;
-};
-
-const filterData = (data, filter) => {
-  const { vstname, group } = filter;
-  return data.filter((item) => {
-    // true || false
-    let result = true;
-
-    if (vstname) {
-      if (item.headline.toLowerCase().indexOf(vstname.toLowerCase()) === -1) {
-        result = false;
-      }
-    }
-
-    if (group) {
-      if (item.group_type !== group) {
-        result = false;
-      }
-    }
-
-    return result;
-  });
-};
-generateContent(PLUGINS);
-document.forms.filtersForm.addEventListener("submit", (e) => {
-  e.preventDefault();
-
-  const formData = new FormData(e.target);
-  const formProps = Object.fromEntries(formData); // {group: STRING || UNDEFINED, name: STRING || UNDEFINED}
-
-  const filteredData = filterData(PLUGINS, formProps);
-  generateContent(filteredData);
-});
-document.getElementById("rstButton").addEventListener("click", () => {
-  generateContent(PLUGINS);
-});
